Posts order request and redirects to order page (needs merging)
